refactor(app): extract mouse mode scanner
